Allow looking up keys in the web ui #264
