+++ app/views/layouts/application.html.erb
@@ -51,5 +51,6 @@
         </a>
       </div>
     </footer>
+    <%= yield :modals %>
   </body>
 </html>
