+++ app/views/lookups/show.html.erb
@@ -1,0 +1,5 @@
+<%= turbo_frame_tag "lookup-result-frame" do %>
+  <code><pre>
+    <%= @result.to_yaml.sub(/^---$/, "") %>
+  </pre></code>
+<% end %>
