+++ test/controllers/lookups_controller_test.rb
@@ -1,0 +1,7 @@
+require "test_helper"
+
+class LookupsControllerTest < ActionDispatch::IntegrationTest
+  # test "the truth" do
+  #   assert true
+  # end
+end
